new changes

+++ routes/web.php
@@ -1,16 +1,18 @@
 <?php
 
-use App\Http\Controllers\UserController;
+use App\Http\Controllers\Admin\UserController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', function () {
-    return view('welcome');
-});
+    return view('admin.layouts.app');
+})->name('admin.dashboard');
 
 // Crud Users
-Route::get('/admin/users', [UserController::class, 'index'])->name('users.index');
-Route::get('/admin/users/create', [UserController::class, 'create'])->name('users.create');
-Route::post('/admin/users/store', [UserController::class, 'store'])->name('users.store');
-Route::get('/admin/users/{user}/edit', [UserController::class, 'edit'])->name('users.edit');
-Route::put('/admin/users/{user}', [UserController::class, 'update'])->name('users.update');
-Route::delete('/admin/users/{user}', [UserController::class, 'destroy'])->name('users.destroy');
+Route::get('/admin/users', [UserController::class, 'index'])->name('admin.users.index');
+Route::get('/admin/users/create', [UserController::class, 'create'])->name('admin.users.create');
+Route::post('/admin/users/store', [UserController::class, 'store'])->name('admin.users.store');
+Route::get('/admin/users/{user}/edit', [UserController::class, 'edit'])->name('admin.users.edit');
+Route::put('/admin/users/{user}', [UserController::class, 'update'])->name('admin.users.update');
+Route::delete('/admin/users/{user}', [UserController::class, 'destroy'])->name('admin.users.destroy');
+
+Route::get('/admin/cuisines', [UserController::class, 'index'])->name('admin.cuisines.index');
